Update current.yml
The Test Tribe - One of the largest software testing communities is organizing the next edition of one of its conferences - "QonfX," in April. The conference is set to happen in San Francisco, Toronto, Calgary, Portugal, Berlin, and Bangalore. It will also be held online.

Kindly add this conference to your list of software testing conferences.